Split test_instrument into test_instrument_function and test_instrument_class
🤖 Created with help from [Claude Code](https://claude.com/claude-code)
